saccos_backend/requirements.txt
django
psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
django-cors-headers
